+++ respawnpopup.js
@@ -4,9 +4,8 @@ class RespawnPopup {
 		this.touchText = 'Tap to respawn';
 		this.otherMediumText = 'Do something to respawn';
 
-		this.pos = createVector(
-			width * (1/2),
-			height * (3/7));
+		this.pos = null;
+		this.reposition();
 		this.textSize = height * (1/19);
 		colorMode(HSB, 255, 255, 255);
 		this.textColor = RespawnPopup.getColorWithRandomHue( 
@@ -86,5 +85,11 @@ class RespawnPopup {
 			random(255),
 			saturation,
 			brightness);
+	}
+
+	reposition() {
+		this.pos = createVector(
+			width * (1/2),
+			height * (3/7));
 	}
 }