+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Player blob slightly bigger than most drone initial size.
 - Larger span of initial drone sizes.
 - Generate random vectors with perlin noise instead of UP/DOWN/LEFT/RIGHT.
+- Refresh on resize.
 
 ## [0.7.0] - 2016-12-19
 ### Added